# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Guardion - Emergency Unlock Ops Platform API
  version: 0.1.0
  description: |
    MVP API for emergency door unlock dispatch.
    No online payments. Accruals are recorded on callcenter accept; invoices are generated for bulk billing later.
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://api.guardion.com
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /cases:
    post:
      tags: [Customer]
      summary: Create a new case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    get:
      tags: [Customer]
      summary: List customer's cases
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CaseStatus"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Cases list
          content:
            application/json:
              schema:
                type: object
                properties:
                  cases:
                    type: array
                    items:
                      $ref: "#/components/schemas/Case"
  /cases/{caseId}:
    get:
      tags: [Customer, Callcenter, Tech]
      summary: Get case detail
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /cases/{caseId}/cancel:
    post:
      tags: [Customer, Callcenter]
      summary: Cancel case
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Case cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{caseId}/evidence/presign:
    post:
      tags: [Customer, Tech]
      summary: Get presigned upload URL
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidencePresignRequest"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidencePresignResponse"
  /cases/{caseId}/evidence/complete:
    post:
      tags: [Customer, Tech]
      summary: Complete evidence upload
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceCompleteRequest"
      responses:
        "201":
          description: Evidence registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
  /cases/{caseId}/otp/request:
    post:
      tags: [Tech]
      summary: Request OTP (sends to customer)
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: OTP issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  case_id:
                    type: string
                    format: uuid
                  expires_at:
                    type: string
                    format: date-time
  /cases/{caseId}/otp/verify:
    post:
      tags: [Customer]
      summary: Verify OTP (completes case)
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                otp:
                  type: string
              required: [otp]
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  case:
                    $ref: "#/components/schemas/Case"
  /cases/{caseId}/opinion:
    post:
      tags: [Tech]
      summary: Submit technical opinion
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitOpinionRequest"
      responses:
        "200":
          description: Opinion submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechOpinionReport"
  /cases/{caseId}/report:
    get:
      tags: [Customer, Callcenter]
      summary: Get report link
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Report link
          content:
            application/json:
              schema:
                type: object
                properties:
                  case_id:
                    type: string
                    format: uuid
                  report_url:
                    type: string
...
  /tech/cases/{caseId}/status:
    post:
      tags: [Tech]
      summary: Update case status
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: "#/components/schemas/TechUpdateableStatus"
                note:
                  type: string
              required: [status]
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
      required: [code, message]
    UserRole:
      type: string
      enum: [customer, callcenter, tech, admin]
    CaseStatus:
      type: string
      enum:
        - SUBMITTED
        - NEED_APPROVAL
        - DISPATCHING
        - ACCEPTED
        - ASSIGNED
        - EN_ROUTE
        - ARRIVED
        - WORKING
        - OTP_PENDING
        - COMPLETED
        - CANCELLED
        - EXPIRED
        - REJECTED
        - FAILED
    TechUpdateableStatus:
      type: string
      enum: [EN_ROUTE, ARRIVED, WORKING, OTP_PENDING]
    EvidenceType:
      type: string
      enum: [door, lock, key, other]
...
    CauseCategory:
      type: string
      enum: [suspected_defect, suspected_user_issue, unknown]
    ConfidenceLevel:
      type: string
      enum: [high, medium, low]
...
    CreateCaseRequest:
      type: object
      properties:
        customer_phone:
          type: string
        address_text:
          type: string
        location_lat:
          type: number
        location_lng:
          type: number
        place_note:
          type: string
        symptom_note:
          type: string
      required: [customer_phone, address_text, location_lat, location_lng]
    Case:
      type: object
      properties:
        case_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CaseStatus"
        risk_level:
          type: integer
        customer_id:
          type: string
          format: uuid
        customer_phone:
          type: string
        address_text:
          type: string
        location_lat:
          type: number
        location_lng:
          type: number
        place_note:
          type: string
        symptom_note:
          type: string
        assigned_org_id:
          type: string
          format: uuid
        assigned_tech_id:
          type: string
          format: uuid
        accepted_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CaseDetail:
      allOf:
        - $ref: "#/components/schemas/Case"
        - type: object
          properties:
            eta_seconds:
              type: integer
            tech_summary:
              type: object
              properties:
                tech_id:
                  type: string
                  format: uuid
                display_name:
                  type: string
                phone_relay:
                  type: string
            evidence:
              type: array
              items:
                $ref: "#/components/schemas/Evidence"
...
    EvidencePresignRequest:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/EvidenceType"
        content_type:
          type: string
        file_name:
          type: string
      required: [type, content_type, file_name]
    EvidencePresignResponse:
      type: object
      properties:
        upload_url:
          type: string
        method:
          type: string
        headers:
          type: object
        object_key:
          type: string
        expires_at:
          type: string
          format: date-time
    EvidenceCompleteRequest:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/EvidenceType"
        object_key:
          type: string
      required: [type, object_key]
    Evidence:
      type: object
      properties:
        evidence_id:
          type: string
          format: uuid
        case_id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/EvidenceType"
        uploader_role:
          $ref: "#/components/schemas/UserRole"
        created_at:
          type: string
          format: date-time
        view_url:
          type: string
...
    SubmitOpinionRequest:
      type: object
      properties:
        cause_category:
          $ref: "#/components/schemas/CauseCategory"
        basis_text:
          type: string
        confidence:
          $ref: "#/components/schemas/ConfidenceLevel"
      required: [cause_category, basis_text, confidence]
    TechOpinionReport:
      type: object
      properties:
        case_id:
          type: string
          format: uuid
        cause:
          $ref: "#/components/schemas/CauseCategory"
        basis_text:
          type: string
        confidence:
          $ref: "#/components/schemas/ConfidenceLevel"
        signed_at:
          type: string
          format: date-time
        disclaimer:
          type: string
...